Added auto-populate feature for database and other stuff
I added a button where you can click it to auto-populate the database. I also reorganized the file structure, so in order to run the program just go to localhost/final/html cause it was getting messy. Last I changed the way the teaches and enrollments table were created, and allowed duplicate ID's (that was my mistake setting that to the primary key in the first place). I set their primary keys instead to auto-increment