Bug fixes w/ enrollments & autopop
Changed enrollmentId to be a concatenated string derived from sstudentid, courseid, semester, and year as opposed to the previous auto-incrementing primary key. Tweaked autopop so it does the same. Also added an ignore into the create view for list_classes because if you clicked twice it threw an error that the view already exists